refactor(caib): consolidate output format handling and formatAge
Extract FormatAge to cmd/caib/common/ using the richer catalog
implementation (months, years, future timestamps). Replace inline
format switching in status.go and catalog/list.go with shared
ResolveOutputFormat + RenderFormatted utilities.
