+++ cmd/caib/catalog/catalog_test.go
@@ -2,7 +2,6 @@ package catalog
 
 import (
 	"testing"
-	"time"
 
 	"github.com/spf13/cobra"
 )
@@ -46,37 +45,6 @@ func TestGetOutputFormat_FallbackWithoutFlag(t *testing.T) {
 	cmd := &cobra.Command{Use: "standalone"}
 	if got := getOutputFormat(cmd); got != testFormatTable {
 		t.Errorf("expected fallback %q, got %q", testFormatTable, got)
-	}
-}
-
-func TestFormatAge(t *testing.T) {
-	tests := []struct {
-		name   string
-		offset time.Duration
-		want   string
-	}{
-		{"seconds ago", 30 * time.Second, "30s"},
-		{"minutes ago", 5 * time.Minute, "5m"},
-		{"hours ago", 3 * time.Hour, "3h"},
-		{"days ago", 2 * 24 * time.Hour, "2d"},
-		{"months ago", 60 * 24 * time.Hour, "2mo"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ts := time.Now().Add(-tt.offset).Format(time.RFC3339)
-			got := formatAge(ts)
-			if got != tt.want {
-				t.Errorf("formatAge(%q) = %q, want %q", ts, got, tt.want)
-			}
-		})
-	}
-}
-
-func TestFormatAge_InvalidTimestamp(t *testing.T) {
-	got := formatAge("not-a-timestamp")
-	if got != "not-a-timestamp" {
-		t.Errorf("expected raw string passthrough, got %q", got)
 	}
 }
